thin out .genaiscript handling
