+++ packages/cli/src/playwright.ts
@@ -5,12 +5,13 @@ import type {
     Page,
 } from "playwright"
 import { TraceOptions } from "../../core/src/trace"
-import { dotGenaiscriptPath, logError, logVerbose } from "../../core/src/util"
+import { logError, logVerbose } from "../../core/src/util"
 import { runtimeHost } from "../../core/src/host"
 import { PLAYWRIGHT_VERSION } from "./version"
 import { ellipseUri } from "../../core/src/url"
 import { PLAYWRIGHT_DEFAULT_BROWSER } from "../../core/src/constants"
 import { ensureDir } from "fs-extra"
+import { getVideoDir } from "../../core/src/workdir"
 
 /**
  * Manages browser instances using Playwright, including launching,
@@ -164,10 +165,7 @@ export class BrowserManager {
         if (incognito || recordVideo) {
             const options = { ...rest } as BrowserContextOptions
             if (recordVideo) {
-                const dir = dotGenaiscriptPath(
-                    "videos",
-                    `${new Date().toISOString().replace(/[:.]/g, "-")}`
-                )
+                const dir = getVideoDir()
                 await ensureDir(dir)
                 trace?.itemValue(`video dir`, dir)
                 options.recordVideo = { dir }
