+++ packages/core/src/transformers.ts
@@ -11,12 +11,13 @@ import type {
 } from "@huggingface/transformers"
 import { NotSupportedError } from "./error"
 import { ChatCompletionMessageParam, ChatCompletionResponse } from "./chattypes"
-import { dotGenaiscriptPath, logVerbose } from "./util"
+import { logVerbose } from "./util"
 import { parseModelIdentifier } from "./models"
 import prettyBytes from "pretty-bytes"
 import { hash } from "./crypto"
 import { PLimitPromiseQueue } from "./concurrency"
 import { deleteUndefinedValues } from "./cleaners"
+import { dotGenaiscriptPath } from "./workdir"
 
 function progressBar(): ProgressCallback {
     const progress: Record<string, number> = {}
